Add 214 ScannerType
