+++ gvm/protocols/gmpv214/__init__.py
@@ -66,11 +66,7 @@ from gvm.protocols.gmpv208.entities.report_formats import (
 )
 from gvm.protocols.gmpv208.entities.results import ResultsMixin
 from gvm.protocols.gmpv208.entities.scan_configs import ScanConfigsMixin
-from gvm.protocols.gmpv208.entities.scanners import (
-    ScannersMixin,
-    ScannerType,
-    get_scanner_type_from_string,
-)
+from gvm.protocols.gmpv208.entities.scanners import ScannersMixin
 from gvm.protocols.gmpv208.entities.secinfo import (
     get_info_type_from_string,
     InfoType,
@@ -88,6 +84,10 @@ from gvm.protocols.gmpv208.gmpv208 import GmpV208Mixin
 # NEW IN 214
 from gvm.protocols.gmpv214.entities.notes import NotesMixin
 from gvm.protocols.gmpv214.entities.overrides import OverridesMixin
+from gvm.protocols.gmpv214.entities.scanners import (
+    ScannerType,
+    get_scanner_type_from_string,
+)
 from gvm.protocols.gmpv214.entities.targets import (
     AliveTest,
     get_alive_test_from_string,
